+++ packages/method/src/utils/appendix.ts
@@ -161,9 +161,10 @@ export class Appendix {
    * ```
    * {
    *  "@context": [
+   *   "https://w3id.org/json-ld-patch/v1",
    *   "https://w3id.org/zcap/v1",
    *   "https://w3id.org/security/data-integrity/v2",
-   *   "https://w3id.org/json-ld-patch/v1"
+   *   "https://btcr2.dev/context/v1"
    *  ],
    *  "patch": [
    *   {